refactor: extract form-layout core CSS to styles folder
