+++ packages/form-layout/src/styles/vaadin-form-layout-styles.d.ts
@@ -6,5 +6,3 @@
 import type { CSSResult } from 'lit';
 
 export const formLayoutStyles: CSSResult;
-
-export const formItemStyles: CSSResult;
