+++ packages/form-layout/src/vaadin-form-item.js
@@ -8,8 +8,8 @@ import { defineCustomElement } from '@vaadin/component-base/src/define.js';
 import { PolylitMixin } from '@vaadin/component-base/src/polylit-mixin.js';
 import { CSSInjectionMixin } from '@vaadin/vaadin-themable-mixin/css-injection-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+import { formItemStyles } from './styles/vaadin-form-item-styles.js';
 import { FormItemMixin } from './vaadin-form-item-mixin.js';
-import { formItemStyles } from './vaadin-form-layout-styles.js';
 
 /**
  * `<vaadin-form-item>` is a Web Component providing labelled form item wrapper
